requirements.txt
pillow~=10.2.0
torch~=2.1.2
transformers~=4.37.0